refactored import_categories & setup_categories to combine function; updated import_categories.sh accordingly, removed setup_categories
