+++ scripts/import_categories.py
@@ -1,13 +1,26 @@
 """
-Category Import Script
+Category Management Script
 
-This script parses the categories.json file and imports the hierarchical 
-category structure into the Supabase categories table.
+This script provides comprehensive category management functionality including:
+1. Import hierarchical category structure from categories.json into Supabase
+2. Populate categories from retailer hierarchy files 
+3. Check category existence and normalization
+4. Debug category normalization for specific products
 
 Usage:
+    # Import from categories.json
     python scripts/import_categories.py
     python scripts/import_categories.py --categories-file data/processed/categories.json
     python scripts/import_categories.py --dry-run
+    
+    # Populate from retailer hierarchy
+    python scripts/import_categories.py --populate-from-hierarchy amazon
+    
+    # Check specific category
+    python scripts/import_categories.py --check-category "Gummy Candies"
+    
+    # Debug normalization
+    python scripts/import_categories.py --debug-normalization "JOLLY RANCHER Gummies"
 """
 
 import os
@@ -27,6 +40,9 @@ load_dotenv(override=True)
 
 # add src to path
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.dirname(__file__)), 'src'))
+
+# import the category normalizer
+from src.crawlers.normalizers.category_normalizer import CategoryNormalizer
 
 # import the category importer class
 class CategoryImporter:    
@@ -171,6 +187,27 @@ class CategoryImporter:
             self.logger.error(f"Error clearing categories: {e}")
             raise
 
+    # populate database with categories from hierarchy file
+    def populate_categories_from_hierarchy(self, retailer: str) -> None:
+        try:
+            hierarchy = load_hierarchy_file(retailer)
+            categories = extract_all_categories(hierarchy)
+            
+            self.logger.info(f"Found {len(categories)} categories in {retailer} hierarchy")
+            
+            created_count = 0
+            for category_name in categories:
+                if not check_category_exists(self.supabase, category_name, self.logger):
+                    # create category using the importer's method
+                    category_id = self.create_category(name=category_name)
+                    if category_id:
+                        created_count += 1
+            
+            self.logger.info(f"Created {created_count} new categories")
+            
+        except Exception as e:
+            self.logger.error(f"Error populating categories: {e}")
+
 # setup logging configuration
 def setup_logging(log_level: str = "INFO") -> None:    
     logging.basicConfig(
@@ -188,6 +225,89 @@ def load_categories_json(file_path: str) -> Dict[str, Any]:
         logging.error(f"Failed to load categories file {file_path}: {e}")
         raise
 
+# load hierarchy file for the given retailer
+def load_hierarchy_file(retailer: str) -> dict:    
+    hierarchy_files = {
+        'amazon': 'data/processed/simplified_amazon.json',
+        'target': 'data/processed/simplified_target.json', 
+        'walmart': 'data/processed/simplified_walmart.json'
+    }
+    
+    file_path = hierarchy_files.get(retailer.lower())
+    if not file_path or not os.path.exists(file_path):
+        # try raw data directory
+        file_path = f'data/raw/{retailer.lower()}_grocery_hierarchy.json'
+        if not os.path.exists(file_path):
+            raise FileNotFoundError(f"Hierarchy file not found for {retailer}")
+    
+    with open(file_path, 'r') as f:
+        return json.load(f)
+
+# extract all category names from hierarchy
+def extract_all_categories(hierarchy: dict) -> list:    
+    categories = []
+    
+    def extract_recursive(items):
+        for item in items:
+            name = item.get('name', '')
+            if name:
+                categories.append(name)
+            sub_items = item.get('sub_items', [])
+            if sub_items:
+                extract_recursive(sub_items)
+    
+    departments = hierarchy.get('departments', [])
+    extract_recursive(departments)
+    return categories
+
+# check if a category exists in the database
+def check_category_exists(supabase, category_name: str, logger):
+    try:
+        # create slug like the normalizer does
+        slug = re.sub(r'[^a-zA-Z0-9\s]', '', category_name.lower())
+        slug = re.sub(r'\s+', '-', slug.strip())
+        
+        result = supabase.table('categories').select('id, name, slug').eq('slug', slug).execute()
+        
+        if result.data:
+            logger.info(f"✓ Category '{category_name}' exists in database with ID: {result.data[0]['id']}")
+            return result.data[0]
+        else:
+            logger.warning(f"✗ Category '{category_name}' (slug: '{slug}') NOT found in database")
+            return None
+    except Exception as e:
+        logger.error(f"Error checking category '{category_name}': {e}")
+        return None
+
+# debug category normalization for a specific product
+def debug_category_normalization(supabase, product_name: str, retailer: str, category: str, logger):
+    try:
+        normalizer = CategoryNormalizer(supabase)
+        
+        logger.info(f"Debugging category normalization for: {product_name}")
+        logger.info(f"Retailer: {retailer}, Raw Category: {category}")
+        
+        # test normalization
+        category_names = normalizer.normalize_category(
+            product_name=product_name,
+            product_url="",
+            retailer_name=retailer,
+            raw_category=category
+        )
+        
+        logger.info(f"Normalized categories: {category_names}")
+        
+        # get category IDs
+        category_ids = normalizer.get_or_create_categories(category_names)
+        logger.info(f"Category IDs: {category_ids}")
+        
+        # check each category in database
+        for cat_name in category_names:
+            check_category_exists(supabase, cat_name, logger)
+            
+    except Exception as e:
+        logger.error(f"Error in debug normalization: {e}")
+
 # create supabase client
 def create_supabase_client():    
     supabase_url = os.getenv("SUPABASE_URL")
@@ -201,11 +321,13 @@ def create_supabase_client():
 # main function
 def main():
     # parse arguments
-    parser = argparse.ArgumentParser(description="Import categories from JSON to Supabase")
+    parser = argparse.ArgumentParser(description="Comprehensive Category Management Tool")
+    
+    # Import categories options
     parser.add_argument(
         "--categories-file", 
         default="data/processed/categories.json",
-        help="Path to categories JSON file"
+        help="Path to categories JSON file for import"
     )
     parser.add_argument(
         "--dry-run", 
@@ -217,6 +339,31 @@ def main():
         action="store_true",
         help="Clear existing categories before import (DANGEROUS!)"
     )
+    
+    # Setup categories options
+    parser.add_argument(
+        "--populate-from-hierarchy", 
+        help="Populate categories from hierarchy file (amazon/target/walmart)"
+    )
+    parser.add_argument(
+        "--check-category", 
+        help="Check if specific category exists in database"
+    )
+    parser.add_argument(
+        "--debug-normalization", 
+        help="Debug category normalization for a product name"
+    )
+    parser.add_argument(
+        "--retailer", 
+        default="amazon", 
+        help="Retailer for debugging (default: amazon)"
+    )
+    parser.add_argument(
+        "--raw-category", 
+        help="Raw category to test with --debug-normalization"
+    )
+    
+    # General options
     parser.add_argument(
         "--log-level", 
         default="INFO",
@@ -231,10 +378,6 @@ def main():
     logger = logging.getLogger(__name__)
     
     try:
-        # load categories data
-        logger.info(f"Loading categories from: {args.categories_file}")
-        categories_data = load_categories_json(args.categories_file)
-        
         # create supabase client
         logger.info("Connecting to Supabase...")
         supabase = create_supabase_client()
@@ -242,24 +385,45 @@ def main():
         # create importer
         importer = CategoryImporter(supabase, dry_run=args.dry_run)
         
-        # clear existing categories if requested
-        if args.clear_existing:
-            if args.dry_run:
-                logger.info("[DRY RUN] Would clear existing categories")
-            else:
-                logger.warning("Clearing existing categories...")
-                importer.clear_existing_categories()
-        
-        # import categories
-        importer.import_categories(categories_data)
-        
-        if args.dry_run:
-            logger.info("Dry run completed. No changes made to database.")
+        # handle different operations
+        if args.check_category:
+            check_category_exists(supabase, args.check_category, logger)
+            
+        elif args.populate_from_hierarchy:
+            importer.populate_categories_from_hierarchy(args.populate_from_hierarchy)
+            
+        elif args.debug_normalization:
+            debug_category_normalization(
+                supabase, 
+                args.debug_normalization, 
+                args.retailer, 
+                args.raw_category or "Gummy Candies",
+                logger
+            )
+            
         else:
-            logger.info("Category import completed successfully!")
+            # default behavior: import from categories file
+            logger.info(f"Loading categories from: {args.categories_file}")
+            categories_data = load_categories_json(args.categories_file)
+            
+            # clear existing categories if requested
+            if args.clear_existing:
+                if args.dry_run:
+                    logger.info("[DRY RUN] Would clear existing categories")
+                else:
+                    logger.warning("Clearing existing categories...")
+                    importer.clear_existing_categories()
+            
+            # import categories
+            importer.import_categories(categories_data)
+            
+            if args.dry_run:
+                logger.info("Dry run completed. No changes made to database.")
+            else:
+                logger.info("Category import completed successfully!")
             
     except Exception as e:
-        logger.error(f"Import failed: {e}")
+        logger.error(f"Operation failed: {e}")
         sys.exit(1)
 
 if __name__ == "__main__":
